review 기능 수정 및 추가

+++ src/main/java/com/ezen/sb/mapper/MovieMapper.java
@@ -12,5 +12,5 @@ public interface MovieMapper {
 	int insertMovie(MovieVO movieModel);
 	int deleteMovie(MovieVO movieModel);
 	
-	MovieVO selectMovie(long movieNum);
+	MovieVO selectMovie(Integer movieNum);
 }
